refactor: Add expo-video plugin for video playback
This commit adds the "expo-video" plugin to the project's dependencies in the package.json file. The plugin is used for video playback functionality in the application. By integrating this plugin, the application will be able to play videos seamlessly and provide an enhanced user experience. This feature improves the multimedia capabilities of the application and allows for the inclusion of video content.