refactor: Add isPaid and price fields to StudyMaterial and Quiz models
This commit adds the "isPaid" and "price" fields to the StudyMaterial and Quiz models. In the "material.js" and "Quiz.js" files, the fields are defined as optional Boolean and Number types respectively. This enhancement allows for the differentiation of paid and free study materials and quizzes, and provides a way to set the price for paid materials and quizzes.