feat: Update image paths and add react-native-razorpay dependency
This commit updates the image paths in the app.json file to use the new SplashScreen.png image. It also adds the "react-native-razorpay" dependency to the package.json file.

These code changes enhance the application by ensuring the correct image is displayed on the splash screen and enabling the integration of the Razorpay payment gateway for future payment functionality.