feat: Update image paths, add react-native-razorpay and react-native-countdown-circle-timer dependencies, and implement lazy loading for images
This commit updates the image paths in the app.json file to use the new SplashScreen.png image. It also adds the "react-native-razorpay" and "react-native-countdown-circle-timer" dependencies to the package.json file. Additionally, it implements lazy loading functionality for images in the PDFViewerScreen component, quiz.details.tsx, and studymaterials/index.tsx.

These code changes enhance the application by ensuring the correct image is displayed on the splash screen, enabling the integration of the Razorpay payment gateway, introducing a countdown timer component, and optimizing page load performance by loading images only when they are visible in the viewport.